moving classes around
