+++ src/test/scala/cognite/spark/v1/fdm/FlexibleDataModelEdgeTest.scala
@@ -2,8 +2,8 @@ package cognite.spark.v1.fdm
 
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
-import cognite.spark.v1.fdm.utils.FDMContainerPropertyTypes
-import cognite.spark.v1.{DefaultSource, FlexibleDataModelRelationFactory, FlexibleDataModelsTestBase, SparkTest}
+import cognite.spark.v1.fdm.utils.{FDMContainerPropertyTypes, FlexibleDataModelTestBase, FlexibleDataModelTestInitializer}
+import cognite.spark.v1.{DefaultSource, FlexibleDataModelRelationFactory, SparkTest}
 import com.cognite.sdk.scala.v1.SpaceCreateDefinition
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyDefinition.EdgeConnection
 import com.cognite.sdk.scala.v1.fdm.common.{DataModelReference, DirectRelationReference, Usage}
@@ -21,7 +21,7 @@ class FlexibleDataModelEdgeTest
     extends FlatSpec
     with Matchers
     with SparkTest
-    with FlexibleDataModelsTestBase {
+    with FlexibleDataModelTestInitializer {
 
   private val startEndNodeContainerExternalId = "sparkDsConnectionsTestContainerStartEndNodes1"
   private val startEndNodeViewExternalId = "sparkDsConnectionsTestViewStartEndNodes1"
